+++ android/src/main/java/com/mapbox/mapboxgl/Convert.java
@@ -256,6 +256,10 @@ class Convert {
       final List logoViewMarginsData = toList(logoViewMargins);
       sink.setLogoViewMargins(toInt(logoViewMarginsData.get(0)), toInt(logoViewMarginsData.get(1)));
     }
+    final Object compassGravity = data.get("compassViewPosition");
+    if(compassGravity != null){
+      sink.setCompassGravity(toInt(compassGravity));
+    }
     final Object compassViewMargins = data.get("compassViewMargins");
     if(compassViewMargins != null){
       final List compassViewMarginsData = toList(compassViewMargins);