+++ example/lib/main.dart
@@ -2,10 +2,14 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:location/location.dart';
+import 'package:mapbox_gl_example/full_map.dart';
+import 'package:mapbox_gl_example/offline_regions.dart';
 
 import 'animate_camera.dart';
+import 'full_map.dart';
 import 'line.dart';
 import 'map_ui.dart';
 import 'move_camera.dart';
@@ -14,24 +18,27 @@ import 'place_circle.dart';
 import 'place_symbol.dart';
 import 'scrolling_map.dart';
 
-final List<Page> _allPages = <Page>[
+final List<ExamplePage> _allPages = <ExamplePage>[
   MapUiPage(),
+  FullMapPage(),
   AnimateCameraPage(),
   MoveCameraPage(),
   PlaceSymbolPage(),
   LinePage(),
   PlaceCirclePage(),
   ScrollingMapPage(),
+  OfflineRegionsPage(),
 ];
 
 class MapsDemo extends StatelessWidget {
-  void _pushPage(BuildContext context, Page page) async {
-    final location = Location();
-    final hasPermissions = await location.hasPermission();
-    if (!hasPermissions) {
-      await location.requestPermission();
+  void _pushPage(BuildContext context, ExamplePage page) async {
+    if (!kIsWeb) {
+      final location = Location();
+      final hasPermissions = await location.hasPermission();
+      if (hasPermissions != PermissionStatus.GRANTED) {
+        await location.requestPermission();
+      }
     }
-
     Navigator.of(context).push(MaterialPageRoute<void>(
         builder: (_) => Scaffold(
               appBar: AppBar(title: Text(page.title)),