+++ example/lib/move_camera.dart
@@ -7,7 +7,7 @@ import 'package:mapbox_gl/mapbox_gl.dart';
 
 import 'page.dart';
 
-class MoveCameraPage extends Page {
+class MoveCameraPage extends ExamplePage {
   MoveCameraPage() : super(const Icon(Icons.map), 'Camera control');
 
   @override
@@ -41,6 +41,7 @@ class MoveCameraState extends State<MoveCamera> {
             height: 200.0,
             child: MapboxMap(
               onMapCreated: _onMapCreated,
+              onCameraIdle: ()=>print("onCameraIdle"),
               initialCameraPosition:
                   const CameraPosition(target: LatLng(0.0, 0.0)),
             ),